Update all dependencies to latest version
Since we're working with sailsjs 0.12 why not update everything else?
Of course a couple of thing brake loose but after a small debug time
everything is back to normal. By the way I added eslint:recommended
rules to tight code style a little more and I'm planning to add
semistandard in the next commit.